+++ src/lib/components/bmapv-circle-layer.vue
@@ -1,0 +1,47 @@
+<script>
+import registerMixin from '../mixins/register-component';
+import pickMixin from '../mixins/pick-mixin';
+const mapvgl = require('mapvgl');
+
+export default {
+  name: 'el-bmapv-circle-layer',
+  mixins: [registerMixin],
+  props: [
+    'type',
+    'color',
+    'size',
+    'radius',
+    'duration',
+    'trial',
+    'random',
+    'data',
+    ...pickMixin.props
+  ],
+  data() {
+    return {
+      converters: {
+      },
+      handlers: {
+        data(value) {
+          this.setData(value);
+        },
+        ...pickMixin.handlers
+      }
+    };
+  },
+  created() {
+  },
+  methods: {
+    __initComponent(options) {
+      this.$bmapComponent = new mapvgl.CircleLayer(options);
+      this.$view.addLayer(this.$bmapComponent);
+    },
+    ...pickMixin.methods
+  },
+  destroyed() {
+    if (this.$bmapComponent && this.$view) {
+      this.$view.removeLayer(this.$bmapComponent);
+    }
+  }
+};
+</script>