+++ src/lib/components/bmapv-polygon-layer.vue
@@ -1,0 +1,47 @@
+<script>
+import registerMixin from '../mixins/register-component';
+import pickMixin from '../mixins/pick-mixin';
+const mapvgl = require('mapvgl');
+
+export default {
+  name: 'el-bmapv-polygon-layer',
+  mixins: [registerMixin],
+  props: [
+    'lineColor',
+    'lineWidth',
+    'lineJoin',
+    'fillColor',
+    'fillOpacity',
+    'data',
+    ...pickMixin.props
+  ],
+  data() {
+    return {
+      propsRedirect: {
+      },
+      converters: {
+      },
+      handlers: {
+        data(value) {
+          this.setData(value);
+        },
+        ...pickMixin.handlers
+      }
+    };
+  },
+  created() {
+  },
+  methods: {
+    __initComponent(options) {
+      this.$bmapComponent = new mapvgl.PolygonLayer(options);
+      this.$view.addLayer(this.$bmapComponent);
+    },
+    ...pickMixin.methods
+  },
+  destroyed() {
+    if (this.$bmapComponent && this.$view) {
+      this.$view.removeLayer(this.$bmapComponent);
+    }
+  }
+};
+</script>