+++ src/lib/components/bmapv-shape-line-layer.vue
@@ -1,0 +1,40 @@
+<script>
+import registerMixin from '../mixins/register-component';
+const mapvgl = require('mapvgl');
+
+export default {
+  name: 'el-bmapv-shape-line-layer',
+  mixins: [registerMixin],
+  props: [
+    'color',
+    'data'
+  ],
+  data() {
+    return {
+      propsRedirect: {
+        effect: 'style'
+      },
+      converters: {
+      },
+      handlers: {
+        data(value) {
+          this.setData(value);
+        }
+      }
+    };
+  },
+  created() {
+  },
+  methods: {
+    __initComponent(options) {
+      this.$bmapComponent = new mapvgl.ShapeLineLayer(options);
+      this.$view.addLayer(this.$bmapComponent);
+    }
+  },
+  destroyed() {
+    if (this.$bmapComponent && this.$view) {
+      this.$view.removeLayer(this.$bmapComponent);
+    }
+  }
+};
+</script>