+++ src/lib/index.js
@@ -10,6 +10,22 @@ import BMapvHeatPointLayer from './components/bmapv-heat-point-layer';
 import BMapvHeatMapLayer from './components/bmapv-heat-map-layer';
 import BMapvHeatGridLayer from './components/bmapv-heat-grid-layer';
 import BMapvSparkLayer from './components/bmapv-spark-layer';
+import BMapvCircleLayer from './components/bmapv-circle-layer';
+import BMapvRippleLayer from './components/bmapv-ripple-layer';
+import BMapvTextLayer from './components/bmapv-text-layer';
+import BMapvFanLayer from './components/bmapv-fan-layer';
+import BMapvSimpleLineLayer from './components/bmapv-simple-line-layer';
+import BMapvLineLayer from './components/bmapv-line-layer';
+import BMapvWallLayer from './components/bmapv-wall-layer';
+import BMapvHeatLineLayer from './components/bmapv-heat-line-layer';
+import BMapvLineFlowLayer from './components/bmapv-line-flow-layer';
+import BMapvLineTripLayer from './components/bmapv-line-trip-layer';
+import BMapvWallTripLayer from './components/bmapv-wall-trip-layer';
+import BMapvFlyLineLayer from './components/bmapv-fly-line-layer';
+import BMapvShapeLayer from './components/bmapv-shape-layer';
+import BMapvShapeLineLayer from './components/bmapv-shape-line-layer';
+import BMapvPolygonLayer from './components/bmapv-polygon-layer';
+import BMapvClusterLayer from './components/bmapv-cluster-layer';
 
 let components = [
   BMapvView,
@@ -19,7 +35,23 @@ let components = [
   BMapvHeatPointLayer,
   BMapvHeatMapLayer,
   BMapvHeatGridLayer,
-  BMapvSparkLayer
+  BMapvSparkLayer,
+  BMapvCircleLayer,
+  BMapvRippleLayer,
+  BMapvTextLayer,
+  BMapvFanLayer,
+  BMapvSimpleLineLayer,
+  BMapvLineLayer,
+  BMapvWallLayer,
+  BMapvHeatLineLayer,
+  BMapvLineFlowLayer,
+  BMapvLineTripLayer,
+  BMapvWallTripLayer,
+  BMapvFlyLineLayer,
+  BMapvShapeLayer,
+  BMapvShapeLineLayer,
+  BMapvPolygonLayer,
+  BMapvClusterLayer
 ];
 
 const install = function(Vue, opts = {}) {